# Set HTTP MaxIdleConnsPerHost to runtime.NumCPU() in benchmarks