This will allow each thread's HTTP connection to be reused when run on multiple processes removing significant benchmarking overhead and making the benchmark less likely to exhaust the number of available sockets.

This results in a significant speedup on my machine:
```
name                  old time/op    new time/op    delta
DeliveryNoReply-16      27.5µs ±11%    15.0µs ± 1%  -45.44%  (p=0.000 n=10+10)
DeliveryWithReply-16    48.4µs ± 4%    25.8µs ± 1%  -46.67%  (p=0.000 n=10+10)

name                  old alloc/op   new alloc/op   delta
DeliveryNoReply-16      16.6kB ± 0%     8.6kB ± 0%  -47.84%  (p=0.000 n=10+9)
DeliveryWithReply-16    33.2kB ± 1%    17.2kB ± 0%  -48.15%  (p=0.000 n=10+10)

name                  old allocs/op  new allocs/op  delta
DeliveryNoReply-16         170 ± 0%       116 ± 0%  -31.92%  (p=0.000 n=10+10)
DeliveryWithReply-16       361 ± 0%       251 ± 0%  -30.53%  (p=0.000 n=10+10)
```

Note that this is not a real speedup, just an improvement to the benchmark.